+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes from ns-3.36 to ns-3.37
 
 * Adds support for channel paging to the **LrWpanPhy** (only placeholder,  a single modulation/band is currently supported).
 * Adds supporting structures used by **LrWpanMac** (PAN descriptor, Command Payload Header, Capability Field).
+* Mac(8|16|48|64)Address address allocation pool is now reset between consecutive runs.
 
 ### Changes to build system